PersonalEconomics/requirements.txt
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.54.1
idna==3.10
iniconfig==2.1.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.0
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.16.0
soupsieve==2.7
stack-data==0.6.3
tinycss2==1.4.0
tkcalendar==1.6.1
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vtk==9.3.1
wcwidth==0.2.13
webencodings==0.5.1
XlsxWriter==3.2.3
yarg==0.1.9
